requirements.txt
# AI and Machine Learning
torch==2.0.1
transformers==4.30.2
detoxify==0.5.1
google-generative-ai==0.3.2

# Data Processing
pandas==2.0.2
numpy==1.24.3